requirements.txt
Django
djangorestframework
django-rest-auth
django-cors-headers
python-dotenv
PyMySQL
pytest
pytest-django
requests
beautifulsoup4

